+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/components/GovukCharacterCountSpec.scala
@@ -15,12 +15,21 @@
  */
 
 package uk.gov.hmrc.govukfrontend.views
-package html
+package components
 
-package object components extends Utils with Aliases {
+import play.twirl.api.HtmlFormat
+import uk.gov.hmrc.govukfrontend.views.viewmodels.charactercount.CharacterCount
+import uk.gov.hmrc.govukfrontend.views.html.components._
+import scala.util.Try
+
+class GovukCharacterCountSpec extends TemplateUnitSpec[CharacterCount, GovukCharacterCount]("govukCharacterCount") {
 
   /**
-    * Top-level implicits for all components
+    * Calls the Twirl template with the given parameters and returns the resulting markup
+    *
+    * @param params
+    * @return [[Try[HtmlFormat.Appendable]]] containing the markup
     */
-  object implicits extends Implicits
+  override def render(params: CharacterCount): Try[HtmlFormat.Appendable] =
+    Try(component(params))
 }